:bug: Fix: admin 정보 리코일 연결 후 수정
